# Optimize and simplify code