Hello, noticed you added a new Nim package, and it looked simple and nice enough for me to try to optimize and simplify it :)


This is the benchmark I've used (it uses a package from nimble called `benchy`), on average it's 2.5 - 5x times faster now, depending on the value and arguments.
```nim
# Compiled with -d:danger
import ../src/millie
import benchy


# On my machine:
# Before:
# normal ............................. 0.828 ms      0.854 ms    ±0.009  x1000
# After:
# normal ............................. 0.330 ms      0.336 ms    ±0.003  x1000
timeIt "normal":
  for x in 0..1000:
    # avg - 0.086 - 0.088, min - 0.083
    var temp = millify(1234567890, precision=3)
    keep temp

# Before:
# teststest .......................... 1.553 ms      1.596 ms    ±0.015  x1000
# After:
# teststest .......................... 0.362 ms      0.373 ms    ±0.005  x1000
timeIt "teststest":
  for x in 0..100:
    doAssert millify(100) == "100"
    doAssert millify(1000) == "1K"
    doAssert millify(1000000) == "1M"
    doAssert millify(1000000000) == "1B"
    doAssert millify(1000000000000) == "1T"

    doAssert millify(-100) == "-100"
    doAssert millify(-1000) == "-1K"
    doAssert millify(-1000000) == "-1M"
    doAssert millify(-1000000000) == "-1B"
    doAssert millify(-1000000000000) == "-1T"

    doAssert millify(100, lowercase=true) == "100"
    doAssert millify(1000, lowercase=true) == "1k"
    doAssert millify(1000000, lowercase=true) == "1m"
    doAssert millify(1000000000, lowercase=true) == "1b"
    doAssert millify(1000000000000, lowercase=true) == "1t"

    doAssert millify(1000, space=true) == "1 K"
    doAssert millify(1000000, space=true) == "1 M"
    doAssert millify(1000000000, space=true) == "1 B"
    doAssert millify(1000000000000, space=true) == "1 T"

    doAssert millify(1234) == "1.2K"
    doAssert millify(1234567) == "1.2M"
    doAssert millify(1234567890, precision=3) == "1.235B"
    doAssert millify(1234566000000, precision=5) == "1.23457T"
``` 

The changes also make the code size (after compilation) for all millie-related code to be smaller ~5 times.

If you feel like I've changed too much, you can only pick the parts that you are willing to change.