@Yardanico 2022-03-30
One thing I'm not sure about is checking if the number is too big - currently in `divideTillEnd` the `denom` is an integer, so with a value big enough it'll fail. If we make `denom` a float as well it'll result in the precision becoming less and less with each iteration.