@bichanna 2022-03-30
Thanks a lot! I'm still very new to Nim. I learned a lot from your comments :)